PacBox: Premultiply sound prom data on load rather than during emulation. Use 16-bit data for 'raw' output instead of int.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@27208 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugins/pacbox/arcade.c
@@ -640,10 +640,10 @@ void renderSprites( unsigned char * buffer )
     }
 }
 
-void playSound( int * buf, int len )
+void playSound( int16_t * buf, int len )
 {
     /* Clear the buffer */
-    memset( buf, 0, sizeof (int)*len);
+    memset( buf, 0, sizeof (int16_t)*len);
 
     /* Exit now if sound is disabled */
     if( (output_devices_ & SoundEnabled) == 0 )
